requirements.txt
fastapi==0.88.0
uvicorn==0.23.2
psycopg2-binary==2.9.7
requests==2.31.0
pyproj==3.6.0
SQLAlchemy==2.0.23
geoalchemy2==0.13.0
python-dotenv==1.0.0
